‘我的’ 修改数据结构

+++ src/main/java/com/fjs/cronus/dto/MineAccountDTO.java
@@ -9,23 +9,22 @@ public class MineAccountDTO extends BaseToStringDTO implements Serializable {
 
     private static final long serialVersionUID = 4978823136905615447L;
 
-    private Integer receivables;
-    private Integer payment;
+    private Double receivables;
+    private Double payment;
 
-    public Integer getReceivables() {
+    public Double getReceivables() {
         return receivables;
     }
 
-    public void setReceivables(Integer receivables) {
+    public void setReceivables(Double receivables) {
         this.receivables = receivables;
     }
 
-    public Integer getPayment() {
+    public Double getPayment() {
         return payment;
     }
 
-    public void setPayment(Integer payment) {
+    public void setPayment(Double payment) {
         this.payment = payment;
     }
-
 }
